Split Draw and Parse into separate functions
